Wed commit

+++ pim-tool-back/src/main/java/vn/elca/training/service/impl/ProjectServiceImpl.java
@@ -8,6 +8,7 @@ import vn.elca.training.repository.ProjectRepository;
 import vn.elca.training.service.ProjectService;
 
 import java.util.List;
+import java.util.Optional;
 
 /**
  * @author vlp
@@ -31,12 +32,39 @@ public class ProjectServiceImpl implements ProjectService {
     }
 
     @Override
-    public Project update(Project project) {
-        return projectRepository.save(project);
+    public Project update(Project project, Project project1) {
+        if (project1 != null) {
+            Optional<Project> checkProject = projectRepository.findById(project1.getId());
+            if (checkProject.isPresent()) {
+                if (checkProject.get().getName().equals(project1.getName()) && checkProject.get().getCustomer().equals(project1.getCustomer()) && checkProject.get().getStatus() == project1.getStatus() && checkProject.get().getStartDate().equals(project1.getStartDate())) {
+                    return projectRepository.save(project);
+                } else {
+                    throw new IllegalArgumentException("A thread has updated this version");
+                }
+            } else {
+                throw new IllegalArgumentException("This project was deleted");
+            }
+        }
+        else {
+//            List<Project> checkProjects = findAll();
+//            for (Project p: checkProjects) {
+//                if (Objects.equals(p.getProjectNumber(), project.getProjectNumber())) {
+//                    throw new IllegalArgumentException("Duplicate project number");
+//                }
+//            }
+            return projectRepository.save(project);
+        }
     }
 
     @Override
     public void remove(Project project) {
-        projectRepository.delete(project);
+        Optional<Project> p = projectRepository.findById(project.getId());
+        if (p.isPresent()) {
+            projectRepository.delete(project);
+        }
+        else {
+            System.out.println("Cannot find any project to delete");
+            throw new IllegalArgumentException("Invalid request! Project not found");
+        }
     }
 }
